Scipy integration changes
