+++ include/fast_matrix_market/read_body_threads.hpp
@@ -61,7 +61,7 @@ namespace fast_matrix_market {
         // Number of concurrent chunks available to work on.
         // Too few may starve workers (such as due to uneven chunk splits)
         // Too many increases costs, such as storing chunk results in memory before they're written.
-        const unsigned inflight_count = 5 * pool.get_num_threads();
+        const unsigned inflight_count = 2 * pool.get_num_threads();
 
         // Start reading chunks and counting lines.
         for (unsigned seed_i = 0; seed_i < inflight_count && instream.good(); ++seed_i) {
