+++ tests/matrices/permissive/permissive_lines_indented.mtx
@@ -1,0 +1,4 @@
+        %%MatrixMarket matrix coordinate real general
+        %comment
+        1 1 1
+        1 1 1
